Availability

+++ kalite/updates/management/commands/videodownload.py
@@ -138,34 +138,39 @@ class Command(UpdatesDynamicCommand, CronCommand):
                 try:
 
                     progress_callback = partial(self.download_progress_callback, video)
-                    try:
-                        # Download via urllib
-                        download_video(video.get("youtube_id"), callback=progress_callback)
 
-                    except URLNotFound:
-                        # Video was not found on amazon cloud service,
-                        #   either due to a KA mistake, or due to the fact
-                        #   that it's a dubbed video.
-                        #
-                        # We can use youtube-dl to get that video!!
-                        logging.debug(_("Retrieving youtube video %(youtube_id)s via youtube-dl") % {"youtube_id": video.get("youtube_id")})
+                    # Don't try to download a file that already exists in the content dir - just say it was successful
+                    # and call it a day!
+                    if not os.path.exists(os.path.join(settings.CONTENT_ROOT, "{id}.mp4".format(id=video.get("youtube_id")))):
 
-                        def youtube_dl_cb(stats, progress_callback, *args, **kwargs):
-                            if stats['status'] == "finished":
-                                percent = 100.
-                            elif stats['status'] == "downloading":
-                                percent = 100. * stats['downloaded_bytes'] / stats['total_bytes']
-                            else:
-                                percent = 0.
-                            progress_callback(percent=percent)
-                        scrape_video(video.get("youtube_id"), quiet=not settings.DEBUG, callback=partial(youtube_dl_cb, progress_callback=progress_callback))
+                        try:
+                            # Download via urllib
+                            download_video(video.get("youtube_id"), callback=progress_callback)
 
-                    except IOError as e:
-                        logging.exception(e)
-                        failed_youtube_ids.append(video.get("youtube_id"))
-                        video_queue.remove_file(video.get("youtube_id"))
-                        time.sleep(10)
-                        continue
+                        except URLNotFound:
+                            # Video was not found on amazon cloud service,
+                            #   either due to a KA mistake, or due to the fact
+                            #   that it's a dubbed video.
+                            #
+                            # We can use youtube-dl to get that video!!
+                            logging.debug(_("Retrieving youtube video %(youtube_id)s via youtube-dl") % {"youtube_id": video.get("youtube_id")})
+
+                            def youtube_dl_cb(stats, progress_callback, *args, **kwargs):
+                                if stats['status'] == "finished":
+                                    percent = 100.
+                                elif stats['status'] == "downloading":
+                                    percent = 100. * stats['downloaded_bytes'] / stats['total_bytes']
+                                else:
+                                    percent = 0.
+                                progress_callback(percent=percent)
+                            scrape_video(video.get("youtube_id"), quiet=not settings.DEBUG, callback=partial(youtube_dl_cb, progress_callback=progress_callback))
+
+                        except IOError as e:
+                            logging.exception(e)
+                            failed_youtube_ids.append(video.get("youtube_id"))
+                            video_queue.remove_file(video.get("youtube_id"))
+                            time.sleep(10)
+                            continue
 
                     # If we got here, we downloaded ... somehow :)
                     handled_youtube_ids.append(video.get("youtube_id"))
